pyproject.toml
[project]
name = "rag-notebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.7",
    "langchain-classic>=1.0.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "pypdf>=6.9.2",
    "langchain-community>=0.4.1",
]
